requirements.txt
# ============================================================================
# Unified Requirements File - All Dependencies with Exact Versions
# ============================================================================
# This file contains all dependencies for the backend, agents, and MCP server.
# Generated from working environment - do not use ranges, use exact versions.
# ============================================================================

# ============================================================================
# Web Framework & API
# ============================================================================
fastapi==0.120.4
uvicorn[standard]==0.38.0
starlette==0.49.3
python-multipart==0.0.12
aiofiles==25.1.0
sse-starlette==3.0.3

# ============================================================================
# Database
# ============================================================================
sqlmodel==0.0.22
SQLAlchemy==2.0.44
sqlalchemy-spanner==1.17.1
psycopg2-binary==2.9.9
alembic==1.17.1
greenlet==3.2.4

# ============================================================================
# Authentication & Security
# ============================================================================
PyJWT==2.10.1
pwdlib==0.2.1
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
Authlib==1.6.5
cryptography==46.0.3
pycparser==2.23
cffi==2.0.0

# ============================================================================
# Configuration & Validation
# ============================================================================
python-dotenv==1.0.1
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
annotated-types==0.7.0
annotated-doc==0.0.3
typing-inspection==0.4.2
typing_extensions==4.15.0
email-validator==2.3.0
dnspython==2.8.0

# ============================================================================
# WebSocket & HTTP
# ============================================================================
websockets==15.0.1
httpx==0.28.1
httpx-sse==0.4.3
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
hyperframe==6.1.0
requests==2.32.5
urllib3==2.5.0
certifi==2025.10.5
idna==3.11

# ============================================================================
# Google AI & Cloud Services
# ============================================================================
google-genai==1.47.0
google-adk==1.17.0
google-cloud-speech==2.34.0
google-cloud-firestore==2.21.0
google-cloud-storage==2.19.0
google-cloud-aiplatform==1.124.0
google-cloud-bigquery==3.38.0
google-cloud-bigtable==2.34.0
google-cloud-discoveryengine==0.13.12
google-cloud-secret-manager==2.25.0
google-cloud-spanner==3.59.0
google-cloud-core==2.5.0
google-cloud-resource-manager==1.15.0
google-cloud-logging==3.12.1
google-cloud-monitoring==2.28.0
google-cloud-trace==1.17.0
google-cloud-appengine-logging==1.7.0
google-cloud-audit-log==0.4.0
google-api-core==2.28.1
google-api-python-client==2.186.0
google-auth==2.42.1
google-auth-httplib2==0.2.1
google-resumable-media==2.7.2
google-crc32c==1.7.1
googleapis-common-protos==1.71.0
grpc-google-iam-v1==0.14.3

# ============================================================================
# Firebase
# ============================================================================
firebase-admin==6.9.0
CacheControl==0.14.3
msgpack==1.1.2

# ============================================================================
# Model Context Protocol (MCP)
# ============================================================================
mcp==1.20.0

# ============================================================================
# AI Agents & Memory
# ============================================================================
mem0ai==1.0.0
openai==2.6.1
posthog==6.7.11
qdrant-client==1.15.1
portalocker==3.2.0

# ============================================================================
# ML Libraries
# ============================================================================
torch==2.9.0
torchaudio==2.9.0
numpy==1.26.4
scipy==1.16.3
sympy==1.14.0
mpmath==1.3.0
filelock==3.20.0
fsspec==2025.10.0
triton==3.5.0

# NVIDIA CUDA Libraries (for PyTorch)
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90

# ============================================================================
# OpenTelemetry & Observability
# ============================================================================
opentelemetry-api==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-exporter-gcp-logging==1.10.0a0
opentelemetry-exporter-gcp-monitoring==1.10.0a0
opentelemetry-exporter-gcp-trace==1.10.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-resourcedetector-gcp==1.10.0a0
opentelemetry-semantic-conventions==0.58b0

# ============================================================================
# gRPC & Protocol Buffers
# ============================================================================
grpcio==1.76.0
grpcio-status==1.71.2
grpc-interceptor==0.15.4
protobuf==5.29.5
proto-plus==1.26.1

# ============================================================================
# Utilities & Helpers
# ============================================================================
click==8.3.0
tenacity==8.5.0
backoff==2.2.1
python-dateutil==2.9.0.post0
pytz==2025.2
tzlocal==5.3.1
watchdog==6.0.0
watchfiles==1.1.1
uvloop==0.22.1
anyio==4.11.0
sniffio==1.3.1
six==1.17.0
pyparsing==3.2.5
uritemplate==4.2.0
charset-normalizer==3.4.4
distro==1.9.0
jiter==0.11.1
tqdm==4.67.1
packaging==25.0
cachetools==6.2.1
absolufy-imports==0.3.1
importlib_metadata==8.7.0
zipp==3.23.0

# ============================================================================
# JSON Schema & Validation
# ============================================================================
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
referencing==0.37.0
rpds-py==0.28.0
attrs==25.4.0

# ============================================================================
# Templates & Markup
# ============================================================================
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
PyYAML==6.0.3

# ============================================================================
# Graph & Data Structures
# ============================================================================
networkx==3.5
shapely==2.1.2
graphviz==0.21

# ============================================================================
# ASN.1 & Security
# ============================================================================
pyasn1==0.6.1
pyasn1_modules==0.4.2
rsa==4.9.1

# ============================================================================
# Documentation & Parsing
# ============================================================================
docstring_parser==0.17.0
sqlparse==0.5.3
cloudpickle==3.1.1
setuptools==80.9.0
